+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- ResizeArray.headAndTail
+
 ## [0.26.0] - 2026-03-07
 ### Changed
 - allow ResizeArray.asArray only on reference types